Make it clear how to use VMamba, Swin etc as an encoder, also removed fixed paths.

+++ export_keypoints.py
@@ -67,13 +67,20 @@ def main():
         config['model'] = yaml.load(f, Loader=yaml.FullLoader)['model']
     # ---------------------------------------------
 
+    #if folder does not exist create it
+    if not os.path.exists(os.path.dirname(args.output_file)):
+        os.makedirs(os.path.dirname(args.output_file))
+
     if os.path.isfile(args.output_file):
         print("Output file already exists : ", args.output_file)
         print("It will be overwritten...")
-        # if input("are you sure? (y/n)") != "y":
-        #     print("Exiting...")
-        #     exit()
-        os.remove(args.output_file)
+        if input("are you sure? (y/n)") != "y":
+            os.remove(args.output_file)
+    
+
+    #if config has self.config['homography_regression_head']['check'] , then make it false
+    if "homography_regression_head" in config["model"].keys():
+        config["model"]["homography_regression_head"]["check"] = False
 
     # dataset
     dataset = getattr(datasets, config['dataset']['type'])(config['dataset'])
@@ -96,9 +103,8 @@ def main():
 
     # dataset
     n=5
-    random_indices_n = [5632] #[6799, 4711,  726, 5632, 7378] #torch.randperm(len(dataset))[:n]
+    random_indices_n = [5632] 
     print(random_indices_n)
-    #subset_dataset = torch.utils.data.Subset(dataset, random_indices_n)
     loader_dataset = torch.utils.data.DataLoader(dataset, batch_size=config['prediction']['batchsize'],
                                                  shuffle=False, num_workers=config['prediction']['num_worker'])
 
@@ -175,7 +181,7 @@ def main():
                     batch, net, config['prediction']['homographic_adaptation'])
             
             prob_ha, prob_o, prob_t = out_dict["out"]["prob"] ,out_dict["out_optical"]["prob"],out_dict["out_thermal"]["prob"]
-            desc_opt,desc_th = out_dict["desc_optical"],out_dict["desc_thermal"] #added by me 
+            desc_opt,desc_th = out_dict["desc_optical"],out_dict["desc_thermal"] 
             # compute the nms probablity
             if config['prediction']['nms'] > 0:
                 if args.single_nms:
@@ -202,48 +208,12 @@ def main():
                         pred = torch.nonzero(
                             (prob.squeeze() > config['prediction']['detection_threshold']).float())
                         
-                        
-                        
-                        # if desc_opt.shape[2:] == prob.shape[2:]:
-                        #     # classic descriptors, directly take values
-                        #     extracted_desc_optical = desc_opt[0,:, pred[:, 0], pred[:, 1]].T
-                        #     extracted_desc_thermal = desc_th[0,:, pred[:, 0], pred[:, 1]].T
-                        # else:
-                        #     H, W = prob.shape[2:]
-                        #     extracted_desc_optical = utils.interpolate_descriptors(pred, desc_opt.squeeze(0), H, W)
-                        #     extracted_desc_thermal = utils.interpolate_descriptors(pred, desc_th.squeeze(0), H, W)
-
-                        # #-------
-                        # import cv2
-                        # import numpy as np
-                        # myopt = np.uint8(batch["optical"]["image"][0][0].cpu().numpy()*255)
-                        # myth = np.uint8(batch["thermal"]["image"][0][0].cpu().numpy()*255)
-                        # predictions = [cv2.KeyPoint(c[1], c[0], 3) for c in pred.cpu().numpy().astype(np.float32)]
-                        # myopt = cv2.drawKeypoints(myopt, predictions,
-                        #                                 outImage=np.array([]),
-                        #                                 color=(0, 0, 255),
-                        #                                 flags=cv2.DRAW_MATCHES_FLAGS_DRAW_RICH_KEYPOINTS)
-                        # myth = cv2.drawKeypoints(myth, predictions,
-                        #                                 outImage=np.array([]),
-                        #                                 color=(0, 0, 255),
-                        #                                 flags=cv2.DRAW_MATCHES_FLAGS_DRAW_RICH_KEYPOINTS)
-                        # cv2.imshow("image",myopt)
-                        # cv2.imshow("image2",myth)
-                        # cv2.waitKey(0)
-                        # #-------
-                        #print(pred.shape)
-
-                        # save the data
-                        #print(pred.shape,extracted_desc_optical.shape)
 
                         output_file.create_group(name)
                         output_file[name].create_dataset(
                             'keypoints', data=pred.cpu().numpy())
                         
-                        # output_file[name].create_dataset(
-                        #     'desc_optical', data=extracted_desc_optical.cpu().numpy()) #added for lightglue
-                        # output_file[name].create_dataset(
-                        #     'desc_thermal', data=extracted_desc_thermal.cpu().numpy()) #added for lightglue
+
 
                         if (args.frequency > 1 and not args.skip_processed):
                             for file in backup_output_files:
@@ -251,10 +221,7 @@ def main():
                                 file[name].create_dataset(
                                     'keypoints', data=pred.cpu().numpy())
                                 
-                                # file[name].create_dataset(
-                                #     'desc_optical', data=extracted_desc_optical.cpu().numpy()) #added for lightglue
-                                # file[name].create_dataset(
-                                #     'desc_thermal', data=extracted_desc_thermal.cpu().numpy()) #added for lightglue
+
 
                             if ((epoch_counter) % args.frequency == 0):
                                 backup_output_files[0].close()
@@ -268,30 +235,11 @@ def main():
                             (prob_t.squeeze() > config['prediction']['detection_threshold']).float())
                         
                         
-                        # #print(pred_o.shape,desc_opt.shape)
-                        # if desc_opt.shape[2:] == prob_o.shape[2:]:
-                        #     # classic descriptors, directly take values
-                        #     extracted_desc_optical = desc_opt[0,:, pred_o[:, 0], pred_o[:, 1]].T
-                        #     extracted_desc_thermal = desc_th[0,:, pred_t[:, 0], pred_t[:, 1]].T
-                        # else:
-                        #     H, W = prob_o.shape[2:]
-                        #     extracted_desc_optical = utils.interpolate_descriptors(pred_o, desc_opt.squeeze(0), H, W)
-                        #     extracted_desc_thermal = utils.interpolate_descriptors(pred_t, desc_th.squeeze(0), H, W)
-
-                        #print(extracted_desc_optical.shape)
-                        # save the data 
-                        #print(pred_o.shape,pred_t.shape)
-                        
                         output_file.create_group(name)
                         output_file[name].create_dataset(
                             'keypoints_optical', data=pred_o.cpu().numpy())
                         output_file[name].create_dataset(
                             'keypoints_thermal', data=pred_t.cpu().numpy())
-                        
-                        # output_file[name].create_dataset(
-                        #     'desc_optical', data=extracted_desc_optical.cpu().numpy()) #added for lightglue
-                        # output_file[name].create_dataset(
-                        #     'desc_thermal', data=extracted_desc_thermal.cpu().numpy()) #added for lightglue
                         
                         
                         if (args.frequency > 1 and not args.skip_processed):
@@ -302,99 +250,11 @@ def main():
                                 file[name].create_dataset(
                                     'keypoints_thermal', data=pred_t.cpu().numpy())
                                 
-                                # file[name].create_dataset(
-                                #     'desc_optical', data=extracted_desc_optical.cpu().numpy()) #added for lightglue
-                                # file[name].create_dataset(
-                                #     'desc_thermal', data=extracted_desc_thermal.cpu().numpy()) #added for lightglue
 
                             if ((epoch_counter) % args.frequency == 0):
                                 backup_output_files[0].close()
                                 backup_output_files.pop(0)
 
-
-            # import matplotlib.pyplot as plt
-            # import numpy as np
-            # import cv2
-
-            # # Prepare optical and thermal images
-            # myopt = np.uint8(batch["optical"]["image"][0][0].cpu().numpy() * 255)
-            # myth = np.uint8(batch["thermal"]["image"][0][0].cpu().numpy() * 255)
-
-            # if config["prediction"]["homographic_adaptation"]['aggregation'] != 'window':
-            #     pred_o = pred
-            #     pred_t = pred
-            # # Create keypoints for optical image
-            # predictions_o = [cv2.KeyPoint(c[1], c[0], 5) for c in pred_o.cpu().numpy().astype(np.float32)]
-            # myopt = cv2.drawKeypoints(
-            #     myopt, 
-            #     predictions_o, 
-            #     outImage=np.array([]), 
-            #     color=(255, 0, 0), 
-            #     flags=cv2.DRAW_MATCHES_FLAGS_DRAW_RICH_KEYPOINTS
-            # )
-
-            # # Create keypoints for thermal image
-            # predictions_t = [cv2.KeyPoint(c[1], c[0], 5) for c in pred_t.cpu().numpy().astype(np.float32)]
-            # myth = cv2.drawKeypoints(
-            #     myth, 
-            #     predictions_t, 
-            #     outImage=np.array([]), 
-            #     color=(255, 0, 0), 
-            #     flags=cv2.DRAW_MATCHES_FLAGS_DRAW_RICH_KEYPOINTS
-            # )
-
-            # # Plot images side by side with matplotlib
-            # fig, axes = plt.subplots(1, 2, figsize=(10, 5))
-
-            # # Display optical image with matches count
-            # axes[0].imshow(myopt, cmap='gray')
-            # axes[0].axis('off')
-            # axes[0].text(
-            #     10, 10, f"#Keypoints: {len(predictions_o)}", 
-            #     color='white', fontsize=12, ha='left', va='top', 
-            #     bbox=dict(boxstyle="round,pad=0.3", edgecolor='white', facecolor='black', alpha=0.7)
-            # )
-
-            # # Display thermal image with matches count
-            # axes[1].imshow(myth, cmap='gray')
-            # axes[1].axis('off')
-            # axes[1].text(
-            #     10, 10, f"#Keypoints: {len(predictions_t)}", 
-            #     color='white', fontsize=12, ha='left', va='top', 
-            #     bbox=dict(boxstyle="round,pad=0.3", edgecolor='white', facecolor='black', alpha=0.7)
-            # )
-
-            # # Adjust layout
-            # plt.tight_layout()
-
-            # # Save the figure
-            # #take path of args.output_file
-            # name_extension = ""
-            # if config["prediction"]["homographic_adaptation"]['aggregation'] == 'window':
-            #     if config["prediction"]["homographic_adaptation"]['weighted_window'] == True:
-            #         name_extension = "weighted_window_{}".format(config["prediction"]["homographic_adaptation"]['window_size'])
-            #     else:
-            #         name_extension = "nonweighted_window_{}".format(config["prediction"]["homographic_adaptation"]['window_size'])
-            # elif config["prediction"]["homographic_adaptation"]['aggregation'] == 'prod':
-            #     name_extension = "prod"
-            #     if config["prediction"]["homographic_adaptation"]['filter_size'] > 0:
-            #         name_extension = "prod_gaussian{}".format(config["prediction"]["homographic_adaptation"]['filter_size'])
-                
-            # path = os.path.dirname(args.output_file)
-            # method_name = args.model_dir.split("/")[-1] + "_"+ name_extension
-            # folder_tosave  = os.path.join(path,method_name)
-            # if not os.path.exists(folder_tosave):
-            #     os.makedirs(folder_tosave)
-            # filtosave = os.path.join( folder_tosave,f"keypoint_comparison_{epoch_counter}.png")
-            # plt.savefig(filtosave, dpi=102, bbox_inches='tight')
-
-            # Display the plot
-            #plt.show()
-    # output_file.close()
-
-    # cv2.waitKey(0)
-    # cv2.destroyAllWindows()
-    # sys.stdout.close()
 
     output_file.close()
